+++ iamf/cli/renderer/audio_element_renderer_base.cc
@@ -43,8 +43,8 @@ absl::StatusOr<size_t> AudioElementRendererBase::RenderLabeledFrame(
 
   size_t num_valid_samples = 0;
   RETURN_IF_NOT_OK(iamf_tools::ArrangeSamplesToRender(
-      labeled_frame, ordered_labels_, kEmptyChannel, samples_to_render_,
-      num_valid_samples));
+      labeled_frame, ordered_labels_, kEmptyChannel, trimming_settings_,
+      samples_to_render_, num_valid_samples));
 
   // Render samples in concrete subclasses.
   current_labeled_frame_ = &labeled_frame;